Setup Production Redux Store
