+++ src/index.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/default*/
 import 'babel-polyfill';
 import React from 'react';
 import {render} from 'react-dom';
