+++ src/store/configureStore.dev.js
@@ -1,0 +1,15 @@
+//function to configure store
+
+import {createStore, applyMiddleware, compose} from 'redux';
+import rootReducer from '../reducers';
+import reduxImmutableStateInvariant from 'redux-immutable-state-invariant';
+
+import thunk from 'redux-thunk';
+
+export default function configureStore(initialState) {
+  return createStore(
+    rootReducer,
+    initialState,
+    applyMiddleware(thunk, reduxImmutableStateInvariant())
+  );
+}
