+++ src/store/configureStore.prod.js
@@ -1,0 +1,14 @@
+//function to configure store
+
+import {createStore, applyMiddleware, compose} from 'redux';
+import rootReducer from '../reducers';
+
+import thunk from 'redux-thunk';
+
+export default function configureStore(initialState) {
+  return createStore(
+    rootReducer,
+    initialState,
+    applyMiddleware(thunk)
+  );
+}
